Fixes specifying VAULT_TOKEN and VAULT_URL on the command line
The command line switches for the vault token and the vault url were not
respecting the values being entered.

When I wrote this code I couldn't find very many examples for properly
specifying an optional argument with the optionparser ruby library, and
it seems to be that the one I ended up using as a template was written
incorrectly. This patch uses the proper syntax, and now those switches
work.